+++ src/Plivo/Resource/MaskingSession/AsyncResponse.cs
@@ -1,0 +1,6 @@
+namespace Plivo.Resource.MaskingSession
+{
+    public class AsyncResponse : BaseResponse
+    {
+    }
+}